requirements.txt
ccxt==4.2.35
flask==3.0.2
python-dotenv==1.0.1
pandas==2.2.2